# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Engepe\u00e7as - Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/productGroups": {
            "get": {
                "tags": [
                    "ProductGroup"
                ],
                "summary": "Obtenha uma lista dos grupos de produtos.",
                "description": "Obtenha todos os grupos de produtos ",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade limite de exibi\u00e7\u00e3o",
                        "required": false,
                        "type": "integer",
                        "default": "15"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "P\u00e1gina a ser exibida",
                        "required": false,
                        "type": "integer",
                        "default": "1"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordena\u00e7\u00e3o do retorno",
                        "required": false,
                        "type": "string",
                        "default": "id"
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Dire\u00e7\u00e3o da ordena\u00e7\u00e3o do retorno",
                        "required": false,
                        "type": "string",
                        "default": "ASC"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Informe a sele\u00e7\u00e3o de campos que devem retornar da consulta separados por virgula",
                        "required": false,
                        "type": "string",
                        "default": "id, hrd_D015_Id, product_utilization_id, name, active"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pesquise por qualquer campo, ao usar este campo as outras consultas ser\u00e3o desconsideradas",
                        "required": false,
                        "type": "string",
                        "default": "Busca"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/ProductGroup"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ProductGroup"
                ],
                "summary": "Armazene um Grupo de Produtos rec\u00e9m-criado no armazenamento.",
                "description": "Grupo de produtos da loja.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Grupo de produtos que deve ser armazenado.",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/ProductGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/ProductGroup"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/productGroups/{id}": {
            "get": {
                "tags": [
                    "ProductGroup"
                ],
                "summary": "Exibir o Grupo de Produtos especificado ",
                "description": "Get ProductGroup",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of ProductGroup",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/ProductGroup"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ProductGroup"
                ],
                "summary": "Atualize o Grupo de Produtos especificado no armazenamento.",
                "description": "Update ProductGroup",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Grupo de Produtos",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Grupo de produtos que deve ser atualizado.",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/ProductGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/ProductGroup"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProductGroup"
                ],
                "summary": "Remova o Grupo de Produtos especificado do armazenamento ",
                "description": "Delete ProductGroup",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of ProductGroup",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ProductGroup": {
            "required": [
                "hrd_D015_Id",
                "product_utilization_id",
                "name",
                "active"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "hrd_D015_Id": {
                    "description": "D015_Id",
                    "type": "integer",
                    "format": "int32"
                },
                "product_utilization_id": {
                    "description": "D015_C008_Id",
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "description": "D015_Nome_Grupo",
                    "type": "string"
                },
                "active": {
                    "description": "D015_Ativo",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "EngepecasAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}